+++ text-based-adventure-i.py
@@ -6,8 +6,8 @@ import setup_classes as sc
 import turn_settup as ts
 
 first_player = sc.Warrior(100, 10, 10, "Player 1")
-first_monster = sc.Warrior(10, 15, 8, "First monster created")
-second_monster = sc.Warrior(20, 15, 8, "Second monster created")
+first_monster = sc.Warrior(10, 15, 9, "First monster created")
+second_monster = sc.Warrior(20, 15, 9, "Second monster created")
 first_item = sc.Item("hp_item", 100, "primer item")
 second_item = sc.Item("atk_item", 100, "primer item")
 third_item = sc.Item("def_item", 100, "primer item")
@@ -16,7 +16,9 @@ first_player.bag_items.append(second_item)
 first_player.bag_items.append(third_item)
 gear_item_1 = sc.Gear(5, -5, "Spear of death")
 gear_item_2 = sc.Gear(-5, 5, "Shield of power")
-
+first_player.add_to_gear_bag(gear_item_1)
+first_player.add_to_gear_bag(gear_item_2)
+#first_player.print_gear_bag()
 #first_player.bag_items_print()
 #first_player.equip_item(first_item)
 #first_player.select_items()